tieba/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
python-decouple==3.8
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0